Fix installer URLs
With b1e83667 (feat: rename cargo-dist => dist, 2024-12-17) the package
name for `cargo-dist` was renamed to `dist`. This causes the
`cargo-dist-schema` package to generate release URLs with
`dist-installer.sh`, and the like, but elsewhere we have hard-coded the
URLs as `cargo-dist-installer.sh`.

Fix the installer URLs to use the correct path.