Submit codesigned macOS binaries for notarization
For macOS's Gatekeeper to fully trust a binary, it must be signed by the
developer and also notarized by Apple.

Start submitting the newly generated binary to Apple for notarization as
part of the macOS codesigning process. This adds two new required
environment variables for codesigning:

- CODESIGN_NOTARIZATION_APPLE_ID: The AppleID of the account that is
  submitting the notarization request.
- CODESIGN_NOTARIZATION_PASSWORD: The app-specific password generated
  for notarization.

Electron has a good guide on how the process works[0] for further
reference.

If we were releasing a pkg installer for macOS we could staple the
notarization to the file, but we're not doing that today, so let's not
bother with adding that feature.

[0] https://github.com/electron/notarize/blob/ad9cdbfeec7c0806cfe46ca10791710f9e1502f3/README.md?plain=1#L30-L41