Don't use axobot for homebrew commits
We don't really want to be using an account we don't own to be listed as
the committer on our homebrew tap.

Given that we're using a PAT, set that user as the committer. This has
the added benefit of making it easier to discover who to contact when
the PAT expires.