Add completion-cmds for shell completion
Add a new `dist.completion-cmds` option which takes a map of binary
names to the subcommand to execute to generate shell completions.

This hooks in very easily with the Homebrew
`generate_completions_from_executable` helper function. We allow three
options for completion generation, `clap-env`, which will use
`clap_complete`'s environment variable option to trigger shell
completion, or `subcommand`, which specifies the subcommand and
parameter format, flag or arg, for the command. All options take an
array of shells.

  ```toml
  [dist.completion-cmds.mybin]
  trigger = "clap-env"
  shells = ["bash","fish"]

  [dist.completion-cmds.otherbin]
  trigger.subcommand = {{ name = "completions", format = "flag" }}
  shells = ["bash","fish","pwsh","zsh"]
  ```