Split generate command across multiple lines
`brew style` fails with an uncorrectable error due to the line length of
our `generate_completions_from_executable` method calls. This isn't a
big issue as the `tap` is still updated, but it creates some noise in
the CI job as it looks like the Homebrew publish job has failed.

Put each argument to `generate_completions_from_executable` on a
separate line to make this easier for `brew style` to manipulate.