Install Homebrew test formula as tap
Homebrew recently began blocking the installation of formulas that are
not in a tap, failing with:

  Error: Homebrew requires formulae to be in a tap, rejecting

To resolve this, we now need to copy our test formula into the actual
Homebrew repository as a tap.

Obviously we need to use a temp directory while doing this, but the
`temp_dir` package only supports using the environment's temp dir, which
will be rejected by Homebrew.

Replace `temp_dir` with the more capable `tempfile`, which allows you to
customize the location of the directory. While we're at it, replace all
usages of `temp_dir` with `tempfile` in this package as its a drop in
replacement.